+++ sdk/resourcemanager/compute/armcompute/options.go
@@ -30,13 +30,6 @@ type AvailabilitySetsClientListAvailableSizesOptions struct {
 	// placeholder for future optional parameters
 }
 
-// AvailabilitySetsClientListBySubscriptionOptions contains the optional parameters for the AvailabilitySetsClient.NewListBySubscriptionPager
-// method.
-type AvailabilitySetsClientListBySubscriptionOptions struct {
-	// The expand expression to apply to the operation. Allowed values are 'instanceView'.
-	Expand *string
-}
-
 // AvailabilitySetsClientListOptions contains the optional parameters for the AvailabilitySetsClient.NewListPager method.
 type AvailabilitySetsClientListOptions struct {
 	// placeholder for future optional parameters
@@ -66,15 +59,6 @@ type CapacityReservationGroupsClientGetOptions struct {
 	// reservations under the capacity reservation group which is a snapshot of the
 	// runtime properties of a capacity reservation that is managed by the platform and can change outside of control plane operations.
 	Expand *CapacityReservationGroupInstanceViewTypes
-}
-
-// CapacityReservationGroupsClientListByResourceGroupOptions contains the optional parameters for the CapacityReservationGroupsClient.NewListByResourceGroupPager
-// method.
-type CapacityReservationGroupsClientListByResourceGroupOptions struct {
-	// The expand expression to apply on the operation. Based on the expand param(s) specified we return Virtual Machine or ScaleSet
-	// VM Instance or both resource Ids which are associated to capacity
-	// reservation group in the response.
-	Expand *ExpandTypesForGetCapacityReservationGroups
 }
 
 // CapacityReservationGroupsClientListBySubscriptionOptions contains the optional parameters for the CapacityReservationGroupsClient.NewListBySubscriptionPager